removed linter errors
