+++ src/curd.js
@@ -1,7 +1,8 @@
-/* eslint-disable import/no-cycle */
-
-import { todoList, populateList } from './index.js';
 import bin from './assets/recyclebin.png';
+
+import options from './assets/options.png';
+
+export const todoList = [];
 
 export const saveTodoListToLocalStorage = () => {
   localStorage.setItem('todoList', JSON.stringify(todoList));
@@ -16,31 +17,6 @@ export const retrieveTodoListFromLocalStorage = () => {
   return [];
 };
 
-export const curdFunctionality = () => {
-  const val = document.getElementById('inputVal');
-
-  const storedValue = localStorage.getItem('inputValue');
-  if (storedValue) {
-    val.value = storedValue;
-  }
-
-  val.addEventListener('keypress', (event) => {
-    if (event.key === 'Enter') {
-      const newItem = {
-        description: val.value,
-        completed: false, // value of new object set to fasle by default
-        index: todoList.length + 1, // Set the index to the next available index
-      };
-
-      todoList.push(newItem);
-
-      alert('hello baby');
-      saveTodoListToLocalStorage();
-      populateList();
-      val.value = '';
-    }
-  });
-};
 export const removeVal = (optionElement, divContainer, x) => {
   optionElement.addEventListener('click', () => {
     optionElement.style.display = 'none';
@@ -63,17 +39,78 @@ export const removeVal = (optionElement, divContainer, x) => {
   });
 };
 
-export function editDescp(x) {
-  const newDescription = prompt(`Enter a new description: ${x}`);
+export const populateList = () => {
+  const parent = document.getElementById('populate');
+  parent.innerHTML = '';
+  const ulList = document.createElement('ul');
+  ulList.classList.add('ulList');
 
-  if (newDescription !== null && newDescription !== '') {
-    // Update the description in the todoList array
-    todoList[x].description = newDescription;
+  for (let x = 0; x < todoList.length; x += 1) {
+    if (todoList[x].completed === false) {
+      const divContainer = document.createElement('div');
+      divContainer.classList.add('divContainer');
+      divContainer.setAttribute('id', `id${x}`);
+      const divSmall = document.createElement('div');
+      divSmall.classList.add('divSmall');
+      const checkbox = document.createElement('input');
+      checkbox.setAttribute('type', 'checkbox');
 
-    // Re-populate the list to reflect the changes
-    populateList();
+      const liList = document.createElement('li');
+      liList.classList.add('liList');
+      liList.setAttribute('id', `descp${x}`);
 
-    // Save the updated todoList to local storage
-    saveTodoListToLocalStorage();
+      liList.textContent = todoList[x].description;
+      liList.addEventListener('click', () => {
+        const newDescription = prompt(`Enter a new description: ${x}`);
+
+        if (newDescription !== null && newDescription !== '') {
+          // Update the description in the todoList array
+          todoList[x].description = newDescription;
+
+          // Re-populate the list
+          populateList();
+
+          // Save the updated todoList to local storage
+          saveTodoListToLocalStorage();
+        }
+      });
+
+      const option = document.createElement('img');
+      option.classList.add('three-dots');
+      option.src = options;
+      removeVal(option, divContainer, x);
+
+      divSmall.appendChild(checkbox);
+      divSmall.appendChild(liList);
+      divContainer.appendChild(divSmall);
+      divContainer.appendChild(option);
+      ulList.appendChild(divContainer);
+    }
   }
-}
+  parent.appendChild(ulList);
+};
+
+export const curdFunctionality = () => {
+  const val = document.getElementById('inputVal');
+
+  const storedValue = localStorage.getItem('inputValue');
+  if (storedValue) {
+    val.value = storedValue;
+  }
+
+  val.addEventListener('keypress', (event) => {
+    if (event.key === 'Enter') {
+      const newItem = {
+        description: val.value,
+        completed: false, // value of new object set to fasle by default
+        index: todoList.length + 1, // Set the index to the next available index
+      };
+
+      todoList.push(newItem);
+
+      saveTodoListToLocalStorage();
+      populateList();
+      val.value = '';
+    }
+  });
+};
